Modified all processes so that they work for any user - each user now has their own database

+++ server/delete_todo.php
@@ -7,23 +7,31 @@ header('Content-Type: application/json; charset=utf8; Accept: application/json')
 $request = file_get_contents('php://input');
 error_log("delete_todo.php - Here is the JSON received: ");
 error_log($request);
-$todo = json_decode($request);
+$todoParam = json_decode($request);
+$todo = $todoParam->todo;
+$uid = $todoParam->uid;
 
-error_log("delete_todo.php - Received data: id=" . $todo->id . ", category=" . $todo->category . ", title=" . $todo->title . ", description=" . $todo->description . ", status=" . $todo->status);
+error_log("delete_todo.php - Received data: uid=" . $uid . ", id=" . $todo->id . ", category=" . $todo->category . ", title=" . $todo->title . ", description=" . $todo->description . ", status=" . $todo->status);
 
-// now insert this mapping
-$db = new SQLite3('db/todo.sqlite');
-try {
-	$statement = $db->prepare('delete from TODO where ID = :id');
-	$statement->bindValue(':id', $todo->id);
-	$statement->execute();
-	$statement->close();
-	$db->close();
-	print_r(json_encode("success"));
-} catch (Exception $e) {
-	error_log("delete_todo.php - Error deleting TODO...  Error message: " . $e->getMessage());
-	$db->close();
-	print_r(json_encode("error"));
+$db = null;
+$filename = 'db/todo_' . $uid . '.sqlite';
+if (file_exists($filename)) {
+	$db = new SQLite3('db/todo_' . $uid . '.sqlite');
+	try {
+		$statement = $db->prepare('delete from TODO where ID = :id');
+		$statement->bindValue(':id', $todo->id);
+		$statement->execute();
+		$statement->close();
+		$db->close();
+		print_r(json_encode("success"));
+	} catch (Exception $e) {
+		error_log("delete_todo.php - Error deleting TODO...  Error message: " . $e->getMessage());
+		$db->close();
+		print_r(json_encode("error"));
+	}
+} else {
+	error_log("delete_todo.php - No database file " . $filename);
+	print_r(json_encode("error|There is no database for user " . $uid));
 }
 
 ?>
